+++ src/main/java/com/ctd/integrador/backend1/persistence/repository/IPacienteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface IPacienteRepository extends JpaRepository<Paciente, Long> {
 
     @Query("SELECT p FROM Paciente p WHERE p.dni = ?1")
     Paciente findPacienteByDni(String dni);
